# server/src/publics/doc/swagger.json
{"swagger":"2.0","info":{"description":"The xmind API description","version":"1.0","title":"my xmind api"},"basePath":"/","tags":[],"schemes":["http"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/":{"get":{"responses":{"200":{"description":""}},"produces":["application/json"],"consumes":["application/json"]}},"/user/role":{"get":{"summary":"获取权限列表","description":"","responses":{"200":{"description":"","schema":{"type":"array","items":{"$ref":"#/definitions/RoleModel"}}}},"tags":["账号权限"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/user/check":{"post":{"summary":"检测 Token","description":"","responses":{"201":{"description":""}},"tags":["账号权限"],"produces":["application/json"],"consumes":["application/json"]}},"/user/signin":{"post":{"summary":"登陆","description":"","parameters":[{"name":"SigninDto","required":true,"in":"body","schema":{"$ref":"#/definitions/SigninDto"}}],"responses":{"201":{"description":""}},"tags":["账号权限"],"produces":["application/json"],"consumes":["application/json"]}},"/user/signup":{"post":{"summary":"注册","description":"","parameters":[{"name":"SignupDto","required":true,"in":"body","schema":{"$ref":"#/definitions/SignupDto"}}],"responses":{"201":{"description":""}},"tags":["账号权限"],"produces":["application/json"],"consumes":["application/json"]}},"/user":{"put":{"summary":"修改用户信息","description":"","parameters":[{"name":"UpdateUserDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateUserDto"}}],"responses":{"200":{"description":""}},"tags":["账号权限"],"produces":["application/json"],"consumes":["application/json"]},"get":{"summary":"获取用户列表","description":"","responses":{"200":{"description":"","schema":{"$ref":"#/definitions/UserModel"}}},"tags":["账号权限"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/user/info":{"get":{"summary":"获取用户信息","description":"","responses":{"200":{"description":"","schema":{"$ref":"#/definitions/UserModel"}}},"tags":["账号权限"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/project":{"post":{"summary":"新建项目","description":"","parameters":[{"name":"AddProjectDto","required":true,"in":"body","schema":{"$ref":"#/definitions/AddProjectDto"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/ProjectDto"}}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]},"get":{"summary":"获取项目列表","description":"","responses":{"200":{"description":""}},"tags":["项目相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/project/{projectId}":{"put":{"summary":"编辑项目","description":"","parameters":[{"type":"number","name":"projectId","required":true,"in":"path"},{"name":"UpdateProjectDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateProjectDto"}}],"responses":{"200":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]},"delete":{"summary":"删除项目","description":"","parameters":[{"type":"number","name":"projectId","required":true,"in":"path"}],"responses":{"200":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/project/info":{"get":{"summary":"获取项目信息","description":"","parameters":[{"type":"number","name":"projectId","required":true,"in":"query"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/ProjectModel"}}},"tags":["项目相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/project/all":{"get":{"summary":"获取所有相关项目","description":"","responses":{"200":{"description":""}},"tags":["项目相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/project/add-members":{"post":{"summary":"添加成员","description":"","parameters":[{"name":"AddMembersDto","required":true,"in":"body","schema":{"$ref":"#/definitions/AddMembersDto"}}],"responses":{"201":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/project/delete-members":{"post":{"summary":"删除成员","description":"","parameters":[{"name":"DeleteMembersDto","required":true,"in":"body","schema":{"$ref":"#/definitions/DeleteMembersDto"}}],"responses":{"201":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/project/update-members":{"post":{"summary":"编辑成员","description":"","parameters":[{"name":"UpdateMembersDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateMembersDto"}}],"responses":{"201":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/project/sourcemap":{"post":{"summary":"添加sourcemap","description":"","parameters":[{"name":"AddSourcemapsDto","required":true,"in":"body","schema":{"$ref":"#/definitions/AddSourcemapsDto"}}],"responses":{"201":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/project/sourcemap/action":{"put":{"summary":"批量操作sourcemap","description":"","parameters":[{"name":"ActionSourcemapsDto","required":true,"in":"body","schema":{"$ref":"#/definitions/ActionSourcemapsDto"}}],"responses":{"200":{"description":""}},"tags":["项目相关"],"produces":["application/json"],"consumes":["application/json"]}},"/search/error.gif":{"post":{"summary":"上报错误日志","description":"","responses":{"200":{"description":""}},"tags":["上传日志"],"produces":["application/json"],"consumes":["application/json"]},"get":{"summary":"上报错误日志","description":"","responses":{"200":{"description":""}},"tags":["上传日志"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/search/log":{"get":{"summary":"查询日志列表","description":"","responses":{"200":{"description":""}},"tags":["上传日志"],"produces":["application/json"],"consumes":["application/json"]}},"/search/stat/error":{"get":{"summary":"统计error","description":"","parameters":[],"responses":{"200":{"description":""}},"tags":["上传日志"],"produces":["application/json"],"consumes":["application/json"]}},"/search/stat/log":{"get":{"summary":"统计error日志","description":"","parameters":[],"responses":{"200":{"description":""}},"tags":["上传日志"],"produces":["application/json"],"consumes":["application/json"]}},"/search/clear":{"get":{"responses":{"200":{"description":""}},"tags":["上传日志"],"produces":["application/json"],"consumes":["application/json"]}},"/common/upload":{"post":{"summary":"上传文件","description":"","responses":{"201":{"description":""}},"tags":["公共"],"produces":["application/json"],"consumes":["application/json"]}},"/error/info":{"get":{"summary":"获取错误信息","description":"","parameters":[{"type":"string","name":"errorId","required":true,"in":"query"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/ErrorModel"}}},"tags":["错误相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/error/all-info":{"get":{"summary":"获取错误信息byAlarm","description":"","parameters":[{"type":"string","name":"errorId","required":true,"in":"query"}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/ErrorModel"}}},"tags":["错误相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/error":{"get":{"summary":"获取错误列表","description":"","responses":{"200":{"description":""}},"tags":["错误相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"put":{"summary":"修改错误","description":"","parameters":[{"name":"UpdateErrorDto","required":true,"in":"body","schema":{"$ref":"#/definitions/UpdateErrorDto"}}],"responses":{"200":{"description":"","schema":{"$ref":"#/definitions/Error"}}},"tags":["错误相关"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"RoleModel":{"type":"object","properties":{"name":{"type":"string"},"code":{"type":"string"},"status":{"type":"number"},"global":{"type":"number"}},"required":["name","code","status","global"]},"SigninDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"SignupDto":{"type":"object","properties":{"username":{"type":"string"},"nickname":{"type":"string"},"password":{"type":"string"}},"required":["username","nickname","password"]},"UpdateUserDto":{"type":"object","properties":{"nickname":{"type":"string"},"email":{"type":"string"},"mobile":{"type":"string"}},"required":["nickname","email","mobile"]},"UserModel":{"type":"object","properties":{}},"AddProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"language":{"type":"string"},"version":{"type":"string"}},"required":["name","description","language","version"]},"ProjectDto":{"type":"object","properties":{}},"UpdateProjectDto":{"type":"object","properties":{"name":{"type":"string"},"description":{"type":"string"},"language":{"type":"string"},"version":{"type":"string"},"frame":{"type":"string"},"sourcemapEnable":{"type":"boolean"},"sourcemapOnline":{"type":"boolean"},"alarmType":{"type":"string"},"alarmHookUrl":{"type":"string"},"guarderId":{"type":"number"}},"required":["name","description","language","version","frame","sourcemapEnable","sourcemapOnline","alarmType","alarmHookUrl","guarderId"]},"ProjectModel":{"type":"object","properties":{}},"AddMembersDto":{"type":"object","properties":{"projectId":{"type":"number"}},"required":["projectId"]},"DeleteMembersDto":{"type":"object","properties":{"projectId":{"type":"number"}},"required":["projectId"]},"UpdateMembersDto":{"type":"object","properties":{"projectId":{"type":"number"}},"required":["projectId"]},"AddSourcemapsDto":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string"}}},"required":["files"]},"ActionSourcemapsDto":{"type":"object","properties":{"sourcemapIds":{"type":"array","items":{"type":"string"}},"actionType":{"type":"string"},"hash":{"type":"boolean"},"version":{"type":"string"}},"required":["sourcemapIds","actionType","hash","version"]},"ErrorModel":{"type":"object","properties":{}},"UpdateErrorDto":{"type":"object","properties":{}},"Error":{"type":"object","properties":{}}}}